Rename api to model
