+++ components/documentation/APINamespace.tsx
@@ -11,7 +11,7 @@ import { FramerAPIContext } from "../contexts/FramerAPIContext"
 import { DeprecatedNotice } from "./DeprecatedNotice"
 import { Grid } from "../layout/Grid"
 import { MissingModelWarning } from "./MissingModelWarning"
-import { NamespaceModel, Kind } from "../../api"
+import { NamespaceModel, Kind } from "../../model"
 import { Permalink } from "../layout/Permalink"
 import { ReleaseBadge } from "./ReleaseBadge"
 import { apiClassName, permalinkId } from "./helpers"
