+++ components/documentation/APIProperty.tsx
@@ -1,5 +1,5 @@
 import * as React from "react"
-import { Kind, PropertyModel } from "../../api"
+import { Kind, PropertyModel } from "../../model"
 import { FramerAPIContext } from "../contexts/FramerAPIContext"
 import { APIOverviewElement } from "./APIOverview"
 import { MissingModelWarning } from "./MissingModelWarning"
