+++ model/__fixtures__/example.data.ts
