+++ model/searchify.ts
